removed ResponseBuilder Interface added Test for Empty String 400 Error Implemented ErrorHandler Interface implemented interface in ServerError
