+++ src/main/java/ErrorHandler.java
@@ -1,0 +1,4 @@
+public interface ErrorHandler {
+
+    Response prepareResponse();
+}
