+++ src/main/java/MethodNotAllowed.java
@@ -1,7 +1,7 @@
 public class MethodNotAllowed {
 
     public static Response prepareResponse(String[] allowedMethods) {
-        TextResponseBuilder responseBuilder = new TextResponseBuilder();
+        ResponseBuilder responseBuilder = new ResponseBuilder();
         Response response = new Response(responseBuilder);
         System.out.println("Method not allowed!");
         responseBuilder.setStatusCode(405);
