+++ src/main/java/MethodOptions2.java
@@ -2,7 +2,7 @@ public class MethodOptions2 implements RoutingInterface{
     String[] allowedMethods = {"GET", "HEAD", "OPTIONS", "PUT", "POST"};
 
     public Response prepareResponse(Request request) {
-        TextResponseBuilder responseBuilder = new TextResponseBuilder();
+        ResponseBuilder responseBuilder = new ResponseBuilder();
         Response response = new Response(responseBuilder);
         System.out.println("Simple Options to Method Options Identified");
         responseBuilder.setStatusCode(200);
