+++ src/main/java/ResourceNotFound.java
@@ -1,7 +1,7 @@
 public class ResourceNotFound implements RoutingInterface{
 
     public Response prepareResponse(Request request) {
-        TextResponseBuilder responseBuilder = new TextResponseBuilder();
+        ResponseBuilder responseBuilder = new ResponseBuilder();
         Response response = new Response(responseBuilder);
         System.out.println("Resource not Found");
         responseBuilder.setStatusCode(404);
